third_party/blink/web_tests/virtual/dir-pseudo-disabled/external/wpt/css/selectors/dir-pseudo-on-bdi-element-expected.txt

This is a testharness.js-based test.
[FAIL] bdi element without dir content attribute
  Failed to execute 'matches' on 'Element': ':dir(ltr)' is not a valid selector.
[FAIL] bdi element with invalid dir content attribute
  Failed to execute 'matches' on 'Element': ':dir(ltr)' is not a valid selector.
[FAIL] bdi element with dir=auto content attribute
  Failed to execute 'matches' on 'Element': ':dir(ltr)' is not a valid selector.
[FAIL] bdi element with dir=ltr content attribute
  Failed to execute 'matches' on 'Element': ':dir(ltr)' is not a valid selector.
[FAIL] bdi element with dir=rtl content attribute
  Failed to execute 'matches' on 'Element': ':dir(ltr)' is not a valid selector.
Harness: the test ran to completion.
